i made a lot of work i dont remember

+++ RestaurantApplicationBackend/src/main/java/com/example/restaurantapplication/repository/IProductRepository.java
@@ -1,9 +1,10 @@
 package com.example.restaurantapplication.repository;
 
+import com.example.restaurantapplication.dto.ProductDTO;
 import com.example.restaurantapplication.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IProductRepository extends JpaRepository<Product, Long>
+public interface IProductRepository extends JpaRepository<Product, Integer>
 {
 
 }
